1:M bidirectional relation between movies and reviews

+++ web_ejem5/src/main/java/es/codeurjc/web/controller/MoviesController.java
@@ -19,6 +19,7 @@ import java.util.List;
 
 import es.codeurjc.web.services.*;
 import es.codeurjc.web.entities.*;
+import es.codeurjc.web.repository.MoviesRepository;
 
 @Controller
 public class MoviesController {
@@ -27,9 +28,6 @@ public class MoviesController {
 
 	@Autowired
 	private MoviesService moviesService;
-	
-	@Autowired
-	private ReviewService reviewService;
 
 	@Autowired
 	private CastService castService;
@@ -37,15 +35,18 @@ public class MoviesController {
     @Autowired
     private ImageService imageService;
 
+	@Autowired
+	private MoviesRepository moviesRepository;
+
     @GetMapping("/movies/{id}")
     public String showMovie(Model model, @PathVariable long id) {
-        Movie movie = moviesService.findById(id);
-        if (movie == null) {
-            return "redirect:/error?status=404";
-        }
-        model.addAttribute("movie", movie);
-		model.addAttribute("movieReviews", reviewService.findByMovie(movie));
-        return "movie_template";
+        Optional<Movie> movie = moviesService.findById(id);
+        if (!movie.isPresent()) {
+            return "movieNotFound_template";
+        } else{
+			model.addAttribute("movie", movie);
+			return "movie_template";
+		}
     }
 
     @GetMapping("/movie/{id}/image")	
@@ -84,22 +85,32 @@ public class MoviesController {
 
     @PostMapping("/movie/{id}/delete")
     public String deleteMovie(Model model, @PathVariable long id) throws IOException {
-        moviesService.deleteById(id);
-        imageService.deleteImage(MOVIES_IMAGES_FOLDER, id);
-        return "movie_deleted_template";
+		Optional<Movie> movie=moviesService.findById(id);
+		if (movie.isPresent()){
+			moviesService.deleteById(id);
+        	imageService.deleteImage(MOVIES_IMAGES_FOLDER, id);
+        	return "movie_deleted_template";
+		} else{
+			return "movieNotFound_template";
+		}
     }
 
 	@GetMapping("/movie/{id}/modify")
 	public String modifyMovieForm(Model model, @PathVariable long id) {
-		Movie movie=moviesService.findById(id);
-		model.addAttribute("movie", movie);
-		model.addAttribute("allCast", castService.findAll());
-		return "modify_movie_template";
+		Optional<Movie> movie=moviesService.findById(id);
+		if (movie.isPresent()){
+			model.addAttribute("movie", movie);
+			model.addAttribute("allCast", castService.findAll());
+			return "modify_movie_template";
+		} else{
+			return "movieNotFound_template";
+		}
 	}
 	
 	@PostMapping("/movie/{id}/modify")
 	public String modifyMovie(Model model, @PathVariable long id, @RequestParam String movieName, @RequestParam String movieArgument,@RequestParam int movieYear,@RequestParam(value = "movieCast", required = false) List<Long> movieCast, @RequestParam String movieTrailer, MultipartFile movieImage) throws IOException {
-		Movie movie=moviesService.findById(id);
+		Movie oldMovie = moviesRepository.findById(id).orElseThrow();
+		Movie updatedMovie;
 		if (movieCast!=null){
 			List<Cast> castList=new ArrayList<Cast>();
 			for (int i=0;i<movieCast.size();i++){
@@ -109,17 +120,15 @@ public class MoviesController {
 					castList.add(cast);
 				}
 			}
-			movie.setCast(castList);
+			updatedMovie=new Movie(movieName, movieArgument, movieYear, castList, movieTrailer);
 		} else{
-			movie.setCast(null);
+			updatedMovie=new Movie(movieName, movieArgument, movieYear, null, movieTrailer);
 		}
-		movie.setArgument(movieArgument);
-		movie.setName(movieName);
-		movie.setYear(movieYear);
-		movie.setTrailer(movieTrailer);
-		
-		imageService.deleteImage(MOVIES_IMAGES_FOLDER, movie.getId());
-		imageService.saveImage(MOVIES_IMAGES_FOLDER, movie.getId(), movieImage);
+		updatedMovie.setId(id);
+		oldMovie.getReviews().forEach(c -> updatedMovie.addReview(c));
+		moviesRepository.save(updatedMovie);
+
+		imageService.saveImage(MOVIES_IMAGES_FOLDER, updatedMovie.getId(), movieImage);
 
 		model.addAttribute("movie", moviesService.findById(id));
